+++ src/main/kotlin/io/extremum/functions/api/url/UrlsHolder.kt
@@ -5,8 +5,6 @@ import org.springframework.stereotype.Component
 
 @Component
 internal class UrlsHolder(
-    @Value("\${extremum.functions.api.consul.uri:}")
-    consulUri: String = "",
     @Value("\${extremum.functions.api.keycloak.uri:}")
     keycloakUri: String = "",
     @Value("\${extremum.functions.api.baseUrl:}")
@@ -15,16 +13,9 @@ internal class UrlsHolder(
     xAppId: String,
 ) {
 
-    val consulUri: String = getConsulUri(consulUri, baseUrl)
-
     val keycloakUri: String = getKeycloakUri(keycloakUri, baseUrl, xAppId)
 
     private companion object {
-        fun getConsulUri(consulUri: String, baseUrl: String): String =
-            consulUri.ifEmpty {
-                baseUrl.replace("://api", "://consul")
-            }
-
         fun getKeycloakUri(keycloakUri: String, baseUrl: String, xAppId: String): String =
             keycloakUri.ifEmpty {
                 baseUrl.replace("://api", "://auth.app-$xAppId")